Test structure refactoring
